--- packages/core/src/calendar/CalendarDate.ts
@@ -1,4 +1,0 @@
-export interface CalendarDate {
-  date: Date
-  disabled?: boolean
-}
