+++ packages/core/src/date-range/DateRangeControlProps.ts
@@ -2,14 +2,14 @@ import {KeyboardEvent, ChangeEventHandler, FocusEventHandler, KeyboardEventHandl
 import {Mask} from '../mask'
 import RenderChild from '../RenderChild'
 import {DateRanges} from './DateRangeControl'
-import {CalendarUtilsProps} from '../calendar'
+import {CalendarUtils} from '../calendar'
 
 export default interface DateRangeControlProps {
   value?: Date | null
   valueTo?: Date | null
   format: string
   isRange?: boolean
-  calendar: CalendarUtilsProps
+  utils: CalendarUtils
   onFocus?: () => void
   onBlur?: () => void
   onKeyDown?: (event: KeyboardEvent) => boolean
