+++ packages/desktop/src/date-picker/DatePicker.md
@@ -5,6 +5,7 @@
       title="Дата"
       format="dd.MM.yyyy"
       value={state.date}
+      utils={new DateFnsCalendarUtils()}
       onChange={(date) => {
         if (date.toString() !== 'Invalid Date') {
           setState({date})
