+++ packages/mobile/src/calendar/Calendar.tsx
@@ -1,5 +1,5 @@
 import React, {FC} from 'react'
-import {Box, Card, CardProps, Grid, Flex, Icon, CalendarControl, CalendarUtils, CalendarUtilsProps, CalendarControlChildrenProps} from '@qiwi/pijma-core'
+import {Box, Card, CardProps, Grid, Flex, Icon, CalendarControl, CalendarUtils, CalendarControlChildrenProps} from '@qiwi/pijma-core'
 import {SelectScroll} from '../select-scroll'
 import {Button} from '../button'
 import {Text} from '../typography'
@@ -9,7 +9,7 @@ export interface CalendarProps {
   dateTo?: Date
   buttonText?: string
   isRange?: boolean
-  calendar?: CalendarUtilsProps
+  utils: CalendarUtils
   onChange?: (date: Date) => void
 }
 
@@ -18,7 +18,7 @@ export const Calendar: FC<CalendarProps> = ({
   dateTo,
   isRange = false,
   buttonText = 'Выбрать',
-  calendar = new CalendarUtils(),
+  utils,
   onChange,
 }) => {
   const getDateItems = (renderProps: CalendarControlChildrenProps) => {
@@ -120,7 +120,7 @@ export const Calendar: FC<CalendarProps> = ({
 
   return (
     <CalendarControl
-      calendar={calendar}
+      utils={utils}
       date={date}
       dateTo={dateTo}
       isRange={isRange}
